+++ resources/src/main/java/com/mad/delivery/resources/CreditCode.java
@@ -2,10 +2,10 @@ package com.mad.delivery.resources;
 
 public class CreditCode {
     public String code;
-    public Integer value;
+    public Double value;
 
 
-    public CreditCode(String code, Integer value) {
+    public CreditCode(String code, Double value) {
         this.code = code;
         this.value = value;
     }
